Continue trying to flesh out exceptions.

+++ Machines/PCCompatible/PIC.hpp
@@ -130,7 +130,7 @@ public:
 		return !awaiting_eoi_ && (requests_ & ~mask_);
 	}
 
-	int acknowledge() {
+	uint8_t acknowledge() {
 		in_service_ = 0x01;
 		int id = 0;
 		while(!(in_service_ & requests_) && in_service_) {
@@ -142,11 +142,11 @@ public:
 			eoi_target_ = id;
 			awaiting_eoi_ = !auto_eoi_;
 			requests_ &= ~in_service_;
-			return vector_base_ + id;
+			return uint8_t(vector_base_ + id);
 		}
 
 		// Spurious interrupt.
-		return vector_base_ + 7;
+		return uint8_t(vector_base_ + 7);
 	}
 
 private:
